requirements.txt
aiogram
python-dotenv
requests
vosk
soundfile
pyaudio
librosa 